requirements/base.txt
defusedxml==0.4.1
Django==1.8.3
django-allauth==0.23.0
django-cors-headers==1.1.0
django-filter==0.10.0
django-rest-auth==0.5.0
djangorestframework==3.1.3
Markdown==2.6.2
oauthlib==1.0.3
psycopg2==2.6.1
Pygments==2.0.2
python3-openid==3.0.7
pytz==2015.7
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.16
Unipath==1.1
wheel==0.24.0
